integrated fluxotaxis
Added script for single and batch runs, and added necessary code to planner class for implementation. Note: Michelle commented out the other navigation strategies and the waypoint function. It might be better to introduce a new navigation type that supports the fixed sensor data that is extracted for DRL. This way they will both work depending on how you initialize the settings file for the simulation.